Crosscheck admin module split
