+++ src/apps/homework/admin/__init__.py
@@ -1,2 +1,9 @@
-import apps.homework.admin.answer.admin  # noqa
-import apps.homework.admin.question.admin  # noqa
+from apps.homework.admin.answer.admin import AnswerAdmin
+from apps.homework.admin.crosscheck.admin import AnswerCrossCheckAdmin
+from apps.homework.admin.question.admin import QuestionAdmin
+
+__all__ = [
+    "AnswerAdmin",
+    "AnswerCrossCheckAdmin",
+    "QuestionAdmin",
+]
